Read topology zone from node
While we are inside k8s, we need to read the topology hint for the
node that our pod is currently running on. In the future we'll
need to be able to override these parameters.
This will require changes to the helm charts, as the containers
now expect a POD_NODE_NAME environment variable to be able to use
TopologyAwareHints.